fossilizers: move evidence inside each fossilizer
