+++ batchfossilizer/batchfossilizer.go
@@ -29,8 +29,8 @@ import (
 	"time"
 
 	log "github.com/sirupsen/logrus"
+	"github.com/stratumn/go-indigocore/batchfossilizer/evidences"
 	"github.com/stratumn/go-indigocore/cs"
-	"github.com/stratumn/go-indigocore/cs/evidences"
 	"github.com/stratumn/go-indigocore/fossilizer"
 	"github.com/stratumn/go-indigocore/monitoring"
 	"github.com/stratumn/go-indigocore/types"
