+++ cmd/couchstore/main.go
@@ -22,8 +22,7 @@ import (
 
 	log "github.com/sirupsen/logrus"
 	"github.com/stratumn/go-indigocore/couchstore"
-	_ "github.com/stratumn/go-indigocore/cs/evidences"
-	_ "github.com/stratumn/go-indigocore/dummyfossilizer"
+	_ "github.com/stratumn/go-indigocore/fossilizer/evidences"
 	"github.com/stratumn/go-indigocore/monitoring"
 	"github.com/stratumn/go-indigocore/store/storehttp"
 	"github.com/stratumn/go-indigocore/utils"
