+++ cmd/dummystore/main.go
@@ -19,9 +19,8 @@ import (
 	"flag"
 
 	log "github.com/sirupsen/logrus"
-	_ "github.com/stratumn/go-indigocore/cs/evidences"
-	_ "github.com/stratumn/go-indigocore/dummyfossilizer"
 	"github.com/stratumn/go-indigocore/dummystore"
+	_ "github.com/stratumn/go-indigocore/fossilizer/evidences"
 	"github.com/stratumn/go-indigocore/monitoring"
 	"github.com/stratumn/go-indigocore/store/storehttp"
 )
