+++ cmd/rethinktmpop/main.go
@@ -18,8 +18,7 @@ package main
 import (
 	"flag"
 
-	_ "github.com/stratumn/go-indigocore/cs/evidences"
-	_ "github.com/stratumn/go-indigocore/dummyfossilizer"
+	_ "github.com/stratumn/go-indigocore/fossilizer/evidences"
 	"github.com/stratumn/go-indigocore/monitoring"
 	"github.com/stratumn/go-indigocore/rethinkstore"
 	"github.com/stratumn/go-indigocore/tendermint"
