+++ cmd/tmstore/main.go
@@ -20,8 +20,7 @@ import (
 	"flag"
 
 	log "github.com/sirupsen/logrus"
-	_ "github.com/stratumn/go-indigocore/cs/evidences"
-	_ "github.com/stratumn/go-indigocore/dummyfossilizer"
+	_ "github.com/stratumn/go-indigocore/fossilizer/evidences"
 	"github.com/stratumn/go-indigocore/monitoring"
 	"github.com/stratumn/go-indigocore/store/storehttp"
 	"github.com/stratumn/go-indigocore/tmstore"
